requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.2.0
billiard==3.3.0.23
celery==3.1.22
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-cors-headers==1.1.0
django-extra-fields==0.5
django-filter==0.11.0
django-phonenumber-field==1.0.0
django-push-notifications==1.4.1
django-suit==0.2.16
django-toolbelt==0.0.1
djangorestframework==3.3.2
ecdsa==0.13
Fabric==1.10.2
gunicorn==19.4.5
holidays==0.4.1
kombu==3.0.34
oauthlib==1.0.3
paramiko==1.16.0
phonenumbers==7.2.5
Pillow==3.0.0
psycopg2==2.6.1
pycrypto==2.6.1
PyJWT==1.4.0
python-dateutil==2.2
python-social-auth==0.2.13
python3-openid==3.0.9
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.6.0
six==1.10.0
static3==0.6.1
